Barely got audio working on iOS and then broke it trying to make it better...

+++ Source/Windows/AudioPlayback.h
@@ -23,7 +23,7 @@ private:
     emscripten::val myAudio;
     bool myHasAudio = false;
     std::string myPath = "";
-    float mySpeed;
+    int mySpeed;
     uint myProgress;
     uint myDuration;
 };
